+++ core/src/test/java/com/asml/apa/wta/core/streams/StreamIntegrationTest.java
@@ -28,8 +28,8 @@ public class StreamIntegrationTest {
 
   @BeforeAll
   static void setUpTmpDirectory() throws IOException {
-    new File("tmp").mkdirs();
-    if (!Files.exists(Path.of("tmp"))) {
+    new File("tmp/wta/streams/serialization/").mkdirs();
+    if (!Files.exists(Path.of("tmp/wta/streams/serialization/"))) {
       throw new IOException();
     }
   }
@@ -202,5 +202,24 @@ public class StreamIntegrationTest {
     for (int i = 0; i <= 20000; i++) {
       assertThat(sutList.get(i)).isEqualTo(i);
     }
+  }
+
+  @Test
+  void serializationFilesActuallyGetGeneratedAndDeleted() throws IOException {
+    Stream<Integer> stream = createSerializingStreamOfNaturalNumbers(10, 10);
+    for (int i = 1; i <= 10; i++) {
+      stream.addToStream(i);
+    }
+    for (int i = 1; i <= 10; i++) {
+      stream.addToStream(i);
+    }
+    for (int i = 1; i <= 10; i++) {
+      stream.addToStream(i);
+    }
+    Path directory = Path.of("tmp/wta/streams/serialization/");
+    long fileCount = Files.list(directory).count();
+    assertThat(fileCount).isEqualTo(3);
+    Stream.deleteAllSerializedFiles();
+    assertThat(!Files.exists(Path.of("tmp/wta/streams/serialization/"))).isTrue();
   }
 }